requirements.txt
mlflow==2.6.0
pandas==1.3.3
scikit-learn==0.24.2
pyngrok==5.1.0
flask==2.0.1
transformers==4.9.2
python-dotenv==0.19.0
